# Conflicts:
#	real simulation/RunParallel/RunSequent_Master.py
#	real simulation/RunParallel/function.py

+++ real simulation/RunParallel/RunSequent2.py
@@ -11,16 +11,22 @@ from define import *
 
         
 def RunTool4Atena(idx):
-    os.remove(pathIdx(idx)+'G7-Cyl-Trial-1_NODES_STRAIN.atf')
-    os.remove(pathIdx(idx)+'G7-Cyl-Trial-1_NODES_STRESS.atf')
+    if os.path.exists(pathIdx(idx)+'G7-Cyl-Trial-1_NODES_STRAIN.atf'):
+        os.remove(pathIdx(idx)+'G7-Cyl-Trial-1_NODES_STRAIN.atf')
+        os.remove(pathIdx(idx)+'G7-Cyl-Trial-1_NODES_STRESS.atf')
     cwd = "C:\\Users\\ADMIN\\Documents\\2.Working-Thinh\\AtenaPool\\"+str(idx)
 
-    subprocess.run("\"C:\\Program Files (x86)\\CervenkaConsulting\\AtenaV5\\AtenaConsole.exe\" \"H:\\02.Working-Thinh\\ATENA-WORKING\\Post.atn\"",
-                cwd="H:\\02.Working-Thinh\\ATENA-WORKING",
-                stdout=subprocess.DEVNULL,
-                shell=True,
-                check=True)
+    # subprocess.run("start /wait cmd /c \"C:\\Program Files (x86)\\CervenkaConsulting\\AtenaV5\\AtenaConsole.exe\" H:\\02.Working-Thinh\\ATENA-WORKING\\Post.atn",
+    #             cwd=cwd,
+    #             stdout=subprocess.DEVNULL,
+    #             shell=True,
+    #             check=True)
 
-# RunTool4Atena()
+    subprocess.run("\"C:\\Program Files (x86)\\CervenkaConsulting\\AtenaV5\\AtenaConsole.exe\" H:\\02.Working-Thinh\\ATENA-WORKING\\Post.atn",
+            cwd=cwd,
+            stdout=subprocess.DEVNULL,
+            shell=True,
+            check=True)
+
